Refactor QuizMenu component and update import statements
